requirements.txt
# This project requires a python3.12 environment
dspy==2.5.43
fuzzywuzzy==0.18.0
ipywidgets==8.1.5
jupyterlab==4.3.4
ollama==0.4.7
pandas==2.2.3
pipdeptree==2.24.0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
textgrad==0.1.6
torch==2.5.1

black==24.10.0
black[jupyter]==24.10.0
coverage==7.6.10
flake8==7.1.1
mypy==1.14.1
pylint==3.3.3
